Turned scheduleNotifications to an async/await function.
It's better to wait for results before saving the collection's data than to save it once before setting up the notifications and again after we have a list of notifications identifiers / after we fail to create notifications. This way creates cleaner code and better user experience (as there are fewer expensive processes happening in succession).