Combined errorHandler and permissionDeniedHandler.
Since we're dealing with both in exactly the same way (set sendNotifications to false and alert the user), there's no need for two parameters. Furthermore, the client (ManageViewController) doesn't care what type of error exactly happened; what matters is the error text, which is shown to the user. The logic for grabbing the text should thus be in the controller dedicated for dealing with notifications, rather than in the VC.